STY: move data padding tests
